review fixes - refactor proj structure and add cleanUp in UserCreationTest
