+++ src/main/java/site/stellarburgers/model/UserDataJson.java
@@ -1,4 +1,4 @@
-package site.stellarburgers;
+package site.stellarburgers.model;
 
 public class UserDataJson {
     private String name;
