+++ src/test/java/site/stellarburgers/ChangeDataTest.java
@@ -5,6 +5,9 @@ import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
 import org.junit.Before;
 import org.junit.Test;
+import site.stellarburgers.client.UserClient;
+import site.stellarburgers.model.User;
+import site.stellarburgers.model.UserCredentials;
 
 import static org.hamcrest.Matchers.equalTo;
